project/requirements.txt
Django==3.2.4
celery==5.1.2
redis==3.5.3
google-api-python-client==2.52.0
djangorestframework==3.13.1